+++ src/main/java/racingcar/domain/Car.java
@@ -1,4 +1,4 @@
-package racingcar.race;
+package racingcar.domain;
 
 import racingcar.util.CommonUtil;
