+++ src/main/java/racingcar/domain/Dice.java
@@ -1,4 +1,4 @@
-package racingcar.race;
+package racingcar.domain;
 
 public interface Dice {
     int generate();
