+++ src/main/java/racingcar/domain/GameResultReceiver.java
@@ -1,4 +1,4 @@
-package racingcar.race;
+package racingcar.domain;
 
 import java.util.List;
