+++ src/main/java/racingcar/domain/Racing.java
@@ -1,4 +1,4 @@
-package racingcar.race;
+package racingcar.domain;
 
 import racingcar.DiceRacingRule;
 import racingcar.RacingDice;
