+++ src/test/java/racingcar/RacingDiceTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import racingcar.race.Dice;
+import racingcar.domain.Dice;
 
 import java.util.ArrayList;
 import java.util.List;
