+++ src/test/java/racingcar/domain/CarsJoinRacingTest.java
@@ -1,4 +1,4 @@
-package racingcar.race;
+package racingcar.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
