+++ src/test/java/racingcar/domain/RacingScoreCardTest.java
@@ -1,4 +1,4 @@
-package racingcar.race;
+package racingcar.domain;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
