+++ src/test/java/racingcar/domain/RacingTestRule.java
@@ -1,4 +1,4 @@
-package racingcar.race;
+package racingcar.domain;
 
 class RacingTestRule implements RacingRule {
